# store/openapi.yaml
openapi: 3.1.0
info:
  title: ProAgentStore API
  description: |
    REST API for ProAgentStore — server-powered AI agents on Cloudflare Workers.
    Agents are backed by Durable Objects; each user gets a personal instance with
    isolated memory, knowledge base, and message history.

    **Auth:** All protected endpoints require a Bearer token in the `Authorization`
    header. Obtain a token via `POST /v1/auth/exchange` (FAS OAuth flow) or
    `POST /v1/auth/github` (direct GitHub OAuth).
  version: 1.0.0
  contact:
    url: https://proagentstore.online
servers:
  - url: https://api.proagentstore.online
    description: Production
security:
  - bearerAuth: []
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: string
    Success:
      type: object
      required: [success]
      properties:
        success:
          type: boolean
          example: true
...
    Agent:
      type: object
      properties:
        id:
          type: string
          format: uuid
        owner_id:
          type: string
        slug:
          type: string
          example: my-support-agent
        name:
          type: string
          example: Support Agent
        description:
          type: string
        category:
          type: string
          example: general
        store_type:
          type: string
          example: agent
        icon:
          type: string
        icon_bg:
          type: string
          example: "#7c3aed"
        model:
          type: string
          example: "@cf/meta/llama-3.2-3b-instruct"
        visibility:
          type: string
          enum: [draft, published]
        status:
          type: string
          enum: [active, inactive]
        cron_schedule:
          type: string
          nullable: true
          example: "0 9 * * *"
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    AgentSummary:
      type: object
      properties:
        id:
          type: string
          format: uuid
        slug:
          type: string
        name:
          type: string
        description:
          type: string
        category:
          type: string
        store_type:
          type: string
        icon:
          type: string
        icon_bg:
          type: string
        model:
          type: string
        status:
          type: string
        creator_login:
          type: string
        creator_avatar:
          type: string
        subscriber_count:
          type: integer
...
    KnowledgeDocument:
      type: object
      required: [title, content]
      properties:
        id:
          type: string
        title:
          type: string
        content:
          type: string
        source:
          type: string
          example: manual
        sourceUrl:
          type: string
          format: uri
...
paths:
...
  /v1/agents/{id}:
    get:
      summary: Get agent
      description: |
        Get a single agent by ID or slug. Published agents are public.
        Draft agents are only visible to the owner or admins.
      operationId: getAgent
      security: []
      tags: [Agents]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          description: Agent UUID or slug
      responses:
        "200":
          description: Agent detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Agent"
        "404":
          description: Agent not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    put:
      summary: Update agent
      description: Update agent metadata. Owner or admin only.
      operationId: updateAgent
      tags: [Agents]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          description: Agent UUID
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                description:
                  type: string
                category:
                  type: string
                icon:
                  type: string
                icon_bg:
                  type: string
                model:
                  type: string
                visibility:
                  type: string
                  enum: [draft, published]
                cron_schedule:
                  type: string
                  nullable: true
      responses:
        "200":
          description: Agent updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Success"
        "400":
          description: Nothing to update
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "403":
          description: Not your agent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Agent not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    delete:
      summary: Delete agent
      description: Permanently delete an agent and all its executions/usage. Owner or admin only.
      operationId: deleteAgent
      tags: [Agents]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          description: Agent UUID
      responses:
        "200":
          description: Agent deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Success"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "403":
          description: Not your agent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Agent not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
  /v1/public/agents/{id}:
    get:
      summary: Public agent detail
      description: Full agent info for the store detail page. No auth required.
      operationId: getPublicAgent
      security: []
      tags: [Public]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          description: Agent UUID or slug
      responses:
        "200":
          description: Public agent data
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                  slug:
                    type: string
                  name:
                    type: string
                  description:
                    type: string
                  category:
                    type: string
                  store_type:
                    type: string
                  icon:
                    type: string
                  icon_bg:
                    type: string
                  model:
                    type: string
                  created_at:
                    type: string
                    format: date-time
                  subscriber_count:
                    type: integer
        "404":
          description: Agent not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /v1/public/agents/{id}/try:
    post:
      summary: Trial chat (no auth)
      description: |
        Try a published agent without creating an account. Uses an ephemeral Durable
        Object keyed by agent + session. Limited to 20 messages per session.
      operationId: tryAgent
      security: []
      tags: [Public]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          description: Agent UUID or slug
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [message]
              properties:
                message:
                  type: string
                sessionId:
                  type: string
                  description: Opaque session identifier from a previous response; omit to start a new trial
      responses:
        "200":
          description: Agent reply with session ID
          content:
            application/json:
              schema:
                type: object
                properties:
                  reply:
                    type: string
                  sessionId:
                    type: string
        "429":
          description: Trial limit reached (20 messages)
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                  sessionId:
                    type: string
  /v1/public/agents/{id}/import-gdoc:
    post:
      summary: Import Google Doc into agent knowledge base
      description: |
        Fetch a publicly shared Google Doc and add its content as a knowledge document.
        Requires a valid PAGS session token (Bearer).
      operationId: importGoogleDoc
      tags: [Public]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [docUrl]
              properties:
                docUrl:
                  type: string
                  format: uri
                  example: "https://docs.google.com/document/d/1BxiMV.../edit"
                  description: URL of a publicly accessible Google Doc
      responses:
        "201":
          description: Document ingested
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/KnowledgeDocument"
        "400":
          description: Invalid URL or inaccessible document
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: Auth required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /v1/public/developers/{login}:
    get:
      summary: Developer public profile
      description: Public creator profile with their published agents.
      operationId: getDeveloperProfile
      security: []
      tags: [Public]
      parameters:
        - name: login
          in: path
          required: true
          schema:
            type: string
          description: GitHub login of the creator
      responses:
        "200":
          description: Developer profile and their agents
          content:
            application/json:
              schema:
                type: object
                properties:
                  developer:
                    type: object
                    properties:
                      login:
                        type: string
                      name:
                        type: string
                      avatar:
                        type: string
                        format: uri
                      bio:
                        type: string
                      website:
                        type: string
                      twitter:
                        type: string
                      roles:
                        type: array
                        items:
                          type: string
                      agentCount:
                        type: integer
                  agents:
                    type: array
                    items:
                      $ref: "#/components/schemas/AgentSummary"
        "404":
          description: Developer not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /v1/public/webhook/{instanceId}/ingest:
    post:
      summary: Webhook knowledge ingestion
      description: |
        Add a document to an instance's knowledge base via webhook.
        Intended for Zapier, Make, n8n, and similar automation tools.
        Auth: Bearer token (the instance owner's PAGS session token).
      operationId: webhookIngest
      tags: [Public]
      parameters:
        - name: instanceId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [title, content]
              properties:
                title:
                  type: string
                content:
                  type: string
                source:
                  type: string
                  default: webhook
                sourceUrl:
                  type: string
                  format: uri
      responses:
        "201":
          description: Document ingested
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/KnowledgeDocument"
        "400":
          description: title and content required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: Bearer token required or invalid
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Instance not found or not active
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"